backend/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
elasticsearch-dsl>=7.4.0
django-elasticsearch-dsl>=7.3.0
psycopg2-binary>=2.9.6
drf-yasg>=1.21.5
python-dotenv>=1.0.0
pyjwt>=2.6.0